generate matrices for rejection level on GPU, variable types, unit test alexnet

+++ compute_matrices_for_rejection_level.py
@@ -2,7 +2,6 @@ import os
 import torch
 from argparse import ArgumentParser, Namespace
 from pathlib import Path
-from multiprocessing import Pool
 from typing import Union
 
 from model_zoo.mlp import MLP
@@ -27,10 +26,10 @@ def parse_args(
     if parser is None:
         parser = ArgumentParser()
     parser.add_argument(
-        "--default_index",
-        type = int,
-        default = 0,
-        help = "Index of default trained networks.",
+        "--experiment_name",
+        type = str,
+        default = None,
+        help = "Name of experiment. Check constants/constants.py.",
     )
     parser.add_argument(
         "--num_samples_rejection_level",
@@ -39,16 +38,16 @@ def parse_args(
         help = "Number of train samples to compute rejection level.",
     )
     parser.add_argument(
-        "--nb_workers",
-        type = int,
-        default = 8,
-        help = "How many processes in parallel for adversarial examples computations.",
-    )
-    parser.add_argument(
         "--temp_dir",
         type = str,
         default = None,
         help = "Temporary directory to save and read data. Useful when using clusters."
+    )
+    parser.add_argument(
+        "--batch_size",
+        type = int,
+        default = 16,
+        help = "Batch size for matrix computation. Number of columns of the matrix processed at the same time on GPU."
     )
     return parser.parse_args()
 
@@ -63,47 +62,50 @@ def compute_one_matrix(args: tuple) -> None:
         label, 
         weights_path, 
         architecture_index, 
-        residual, 
-        input_shape, 
+        input_shape,
         num_classes,
-        default_index, 
-        dropout, 
-        i, 
-        temp_dir
+        experiment_name,
+        i,
+        temp_dir,
+        device,
+        batch_size
     ) = args
+
+    im = im.to(device)
+    label = label.to(device)
 
     model = get_model(
         path = weights_path,
         architecture_index = architecture_index,
-        residual = residual,
         input_shape = input_shape,
         num_classes = num_classes,
-        dropout = dropout
+        device = device
     )
     if isinstance(model, MLP):
         matrix_computer = MlpRepresentation(model)
     elif isinstance(model, (CNN_2D, AlexNet, VGG, ResNet)):
-        matrix_computer = ConvRepresentation_2D(model, batch_size=16)
+        matrix_computer = ConvRepresentation_2D(model, batch_size=batch_size, device=device)
     else:
         raise NotImplementedError(f"Model {type(model)} not supported")
-    pred = torch.argmax(model.forward(im))
+
     if temp_dir is not None:
-        path_experiment_matrix = Path(f'{temp_dir}/experiments/{default_index}/rejection_levels/matrices/{i}/matrix.pth')
+        path_experiment_matrix = Path(f'{temp_dir}/experiments/{experiment_name}/rejection_levels/matrices/{i}/matrix.pth')
+        path_prediction = Path(f'{temp_dir}/experiments/{experiment_name}/rejection_levels/matrices/{i}/prediction.pth')
+        Path(f'{temp_dir}/experiments/{experiment_name}/rejection_levels/matrices/{i}/').mkdir(parents=True, exist_ok=True)
     else:
-        path_experiment_matrix = Path(f'experiments/{default_index}/rejection_levels/matrices/{i}/matrix.pth')
+        path_experiment_matrix = Path(f'experiments/{experiment_name}/rejection_levels/matrices/{i}/matrix.pth')
+        path_prediction = Path(f'experiments/{experiment_name}/rejection_levels/matrices/{i}/prediction.pth')
+        Path(f'experiments/{experiment_name}/rejection_levels/matrices/{i}/').mkdir(parents=True, exist_ok=True)
+
+    pred = torch.argmax(model.forward(im))
     # if it is not correctly classified, do not use it for rejection level
     if pred != label:
         return
 
     if os.path.exists(path_experiment_matrix):
         return
+
     mat = matrix_computer.forward(im)
-    if temp_dir is not None:
-        path_prediction = Path(f'{temp_dir}/experiments/{default_index}/rejection_levels/matrices/{i}/prediction.pth')
-        Path(f'{temp_dir}/experiments/{default_index}/rejection_levels/matrices/{i}/').mkdir(parents=True, exist_ok=True)
-    else:
-        path_prediction = Path(f'experiments/{default_index}/rejection_levels/matrices/{i}/prediction.pth')
-        Path(f'experiments/{default_index}/rejection_levels/matrices/{i}/').mkdir(parents=True, exist_ok=True)
 
     torch.save(pred, path_prediction)
     torch.save(mat, path_experiment_matrix)
@@ -112,15 +114,14 @@ def compute_one_matrix(args: tuple) -> None:
 def compute_matrices_for_rejection_level(
         exp_dataset_train: torch.Tensor,
         exp_dataset_labels: torch.Tensor,
-        default_index: int,
-        weights_path: str,
+        experiment_name: str,
+        weights_path: Path,
         architecture_index: int,
-        residual: bool,
-        input_shape: tuple[int,int,int],
+        input_shape,
         num_classes: int,
-        dropout: bool,
-        nb_workers: int = 8,
-        temp_dir:Union[str, None] = None
+        temp_dir:Union[str, None] = None,
+        device: str = 'cpu',
+        batch_size: int = 16,
     ) -> None:
     """
         Args:
@@ -135,22 +136,23 @@ def compute_matrices_for_rejection_level(
             nb_workers: the number of workers.
             temp_dir: the temporary directory.
     """
-    Path(f'experiments/{default_index}/rejection_levels/').mkdir(parents=True, exist_ok=True)
+    Path(f'experiments/{experiment_name}/rejection_levels/').mkdir(parents=True, exist_ok=True)
     print("Computing matrices for rejection level...", flush=True)
 
-    with Pool(processes=nb_workers) as pool:
-        args = [(exp_dataset_train[i],
-                 exp_dataset_labels[i],
-                 weights_path,
-                 architecture_index,
-                 residual,
-                 input_shape,
-                 num_classes,
-                 default_index,
-                 dropout,
-                 i,
-                 temp_dir) for i in range(len(exp_dataset_train))]
-        pool.map(compute_one_matrix, args)
+    for i in range(len(exp_dataset_train)):
+        args = (exp_dataset_train[i],
+                exp_dataset_labels[i],
+                weights_path,
+                architecture_index,
+                input_shape,
+                num_classes,
+                experiment_name,
+                i,
+                temp_dir,
+                device,
+                batch_size)
+
+        compute_one_matrix(args)
 
 
 def main() -> None:
@@ -158,63 +160,66 @@ def main() -> None:
         Main function to compute the matrices for the rejection level.
     """
     args = parse_args()
-    if args.default_index is not None:
-        try:
-            experiment = DEFAULT_EXPERIMENTS[f'experiment_{args.default_index}']
-
-            architecture_index = experiment['architecture_index']
-            residual = experiment['residual']
-            dropout = experiment['dropout']
-            dataset = experiment['dataset']
-            epoch = experiment['epoch'] - 1
-
-        except KeyError:
-            print(f"Error: Default index {args.default_index} does not exist.")
-            return -1
+    if args.experiment_name is not None:
+        experiment = DEFAULT_EXPERIMENTS[f'{args.experiment_name}']
+        architecture_index = experiment['architecture_index']
+        dataset = experiment['dataset']
+        epoch = experiment['epoch'] - 1
 
     else:
         raise ValueError("Default index not specified in constants/constants.py")
 
-    print("Computing matrices for rejection level for Experiment: ", args.default_index,flush=True)
+    print("Computing matrices for rejection level for Experiment: ", args.experiment_name,flush=True)
 
     if args.temp_dir is not None:
-        weights_path = Path(f'{args.temp_dir}/experiments/{args.default_index}/weights') / f'epoch_{epoch}.pth'
+        weights_path = Path(f'{args.temp_dir}/experiments/{args.experiment_name}/weights') / f'epoch_{epoch}.pth'
     else:
-        weights_path = Path(f'experiments/{args.default_index}/weights') / f'epoch_{epoch}.pth'
+        weights_path = Path(f'experiments/{args.experiment_name}/weights') / f'epoch_{epoch}.pth'
 
     if not weights_path.exists():
         raise ValueError(f"Experiment needs to be trained")
 
     input_shape = get_input_shape(dataset)
     num_classes = get_num_classes(dataset)
-    train_set, _ = get_dataset(
-        data_set = dataset,
-        data_loader = False
-    )
-    exp_dataset_train, exp_dataset_labels = subset(
-        train_set = train_set,
-        length = args.num_samples_rejection_level,
-        input_shape = input_shape
-    )
 
-    Path(f'experiments/{args.default_index}/rejection_levels/').mkdir(parents=True, exist_ok=True)
-    torch.save(
-        obj = exp_dataset_train, 
-        f = f'experiments/{args.default_index}/rejection_levels/exp_dataset_train.pth'
-    )
+    Path(f'experiments/{args.experiment_name}/rejection_levels/').mkdir(parents=True, exist_ok=True)
+    exp_dataset_train_file = Path(f'experiments/{args.experiment_name}/rejection_levels/exp_dataset_train.pth')
+    exp_dataset_labels_file = Path(f'experiments/{args.experiment_name}/rejection_levels/exp_dataset_labels.pth')
+
+    if exp_dataset_train_file.exists():
+        exp_dataset_train = torch.load(exp_dataset_train_file)
+        exp_dataset_labels = torch.load(exp_dataset_labels_file)
+
+    else:
+        train_set, _ = get_dataset(
+            data_set=dataset,
+            data_loader=False
+        )
+        exp_dataset_train, exp_dataset_labels = subset(
+            train_set = train_set,
+            length = args.num_samples_rejection_level,
+            input_shape = input_shape
+        )
+
+        torch.save(
+            obj = exp_dataset_train,
+            f = f'experiments/{args.experiment_name}/rejection_levels/exp_dataset_train.pth'
+        )
+        torch.save(
+            obj = exp_dataset_labels,
+            f = f''
+        )
 
     compute_matrices_for_rejection_level(
         exp_dataset_train = exp_dataset_train,
         exp_dataset_labels = exp_dataset_labels,
-        default_index = args.default_index,
+        experiment_name = args.experiment_name,
         weights_path = weights_path,
         architecture_index = architecture_index,
-        residual = residual,
         input_shape = input_shape,
         num_classes = num_classes,
-        dropout = dropout,
-        nb_workers = args.nb_workers,
-        temp_dir = args.temp_dir
+        temp_dir = args.temp_dir,
+        batch_size = args.batch_size
     )
 
     if args.temp_dir is not None:
